+++ executor/pkg_test.go
@@ -129,7 +129,7 @@ func prepareOneColChildExec(sctx sessionctx.Context, rowCount int) Executor {
 	return exec
 }
 
-func buildExec4RadixHashJoin(sctx sessionctx.Context, rowCount int) *HashJoinExec {
+func buildExec4RadixHashJoin(sctx sessionctx.Context, rowCount int) *RadixHashJoinExec {
 	childExec0 := prepareOneColChildExec(sctx, rowCount)
 	childExec1 := prepareOneColChildExec(sctx, rowCount)
 
@@ -148,7 +148,7 @@ func buildExec4RadixHashJoin(sctx sessionctx.Context, rowCount int) *HashJoinExe
 		innerExec:      childExec0,
 		outerExec:      childExec1,
 	}
-	return hashJoinExec
+	return &RadixHashJoinExec{HashJoinExec: hashJoinExec}
 }
 
 func (s *pkgTestSuite) TestRadixPartition(c *C) {
@@ -279,7 +279,7 @@ func (s *pkgTestSuite) TestParallelBuildHashTable4RadixJoin(c *C) {
 	err := hashJoinExec.Open(ctx)
 	c.Assert(err, IsNil)
 
-	hashJoinExec.fetchInnerAndBuildHashTable(ctx)
+	hashJoinExec.partitionInnerAndBuildHashTables(ctx)
 	innerParts := hashJoinExec.innerParts
 	c.Assert(len(hashJoinExec.hashTables), Equals, len(innerParts))
 	for i := 0; i < len(innerParts); i++ {
